// src/types/polyfills.ts
export interface PolyfillLib {
  aptPackage?: string;
  command?: string;
  default: boolean;
  needs: string[];
  path?: string;
}

export type PolyfillKey =
  | 'acl'
  | 'ant-optional'
  | 'ant'
  | 'aria2'
  | 'autoconf'
  | 'automake'
  | 'binutils'
  | 'bison'
  | 'brotli'
  | 'build-essential'
  | 'bzip2'
  | 'ca-certificates'
  | 'coreutils'
  | 'curl'
  | 'dbus'
  | 'dnsutils'
  | 'dpkg-dev'
  | 'dpkg'
  | 'fakeroot'
  | 'file'
  | 'flex'
  | 'fonts-noto-color-emoji'
  | 'ftp'
  | 'g++'
  | 'gcc'
  | 'gnupg'
  | 'gnupg2'
  | 'haveged'
  | 'imagemagick'
  | 'iproute2'
  | 'iputils-ping'
  | 'jq'
  | 'kmod'
  | 'lib32z1'
  | 'libcurl4'
  | 'libffi-dev'
  | 'libgbm-dev'
  | 'libgconf-2-4'
  | 'libgsl-dev'
  | 'libmagic-dev'
  | 'libmagickcore-dev'
  | 'libmagickwand-dev'
  | 'libsecret-1-dev'
  | 'libsqlite3-dev'
  | 'libssl-dev'
  | 'libtool'
  | 'libunwind8'
  | 'libxkbfile-dev'
  | 'libxss1'
  | 'libyaml-dev'
  | 'locales'
  | 'lz4'
  | 'm4'
  | 'make'
  | 'mediainfo'
  | 'mercurial'
  | 'net-tools'
  | 'netcat'
  | 'nodejs'
  | 'openjdk-17-jre-headless"'
  | 'openssh-client'
  | 'p7zip-full'
  | 'p7zip-rar'
  | 'parallel'
  | 'pass'
  | 'patchelf'
  | 'pigz'
  | 'pkg-config'
  | 'pollinate'
  | 'python-is-python3'
  | 'python3-dev'
  | 'python3-pip'
  | 'python3-venv'
  | 'python3'
  | 'rpm'
  | 'rsync'
  | 'shellcheck'
  | 'sphinxsearch'
  | 'sqlite3'
  | 'ssh'
  | 'sshpass'
  | 'subversion'
  | 'sudo'
  | 'swig'
  | 'tar'
  | 'telnet'
  | 'texinfo'
  | 'time'
  | 'tk'
  | 'tzdata'
  | 'uidmap'
  | 'unzip'
  | 'upx'
  | 'wget'
  | 'xorriso'
  | 'xvfb'
  | 'xz-utils'
  | 'yarn'
  | 'zip'
  | 'zsync';